% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort.r
\name{cohort}
\alias{cohort}
\title{Cohort a dataset.}
\usage{
cohort(
  x,
  on,
  name = "data",
  remove_equiv_columns = FALSE,
  keep_cols = character(),
  verbose = FALSE
)
}
\arguments{
\item{x}{a data.frame.}

\item{on}{which variable should be collapsed?}

\item{name}{the variable name of the resulting embedded data.frames.}

\item{remove_equiv_columns}{should equivalent columns be removed?}

\item{keep_cols}{column names that should be kept.}

\item{verbose}{should information about dropped columns be printed?
(default FALSE)}
}
\value{
The collapsed data.frame with numerically encoded columens removed.
}
\description{
Cohort a dataset.
}
