% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consolidate.r
\name{consolidate}
\alias{consolidate}
\title{Consolidate a set of similarly-cohorted data sets}
\usage{
consolidate(data_list, on)
}
\arguments{
\item{data_list}{the list of similarly-cohorted data sets.}

\item{on}{the variable to consolidate on.}
}
\description{
Consolidate a set of similarly-cohorted data sets
}
