% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe.r
\name{consolidated_describe_data}
\alias{consolidated_describe_data}
\title{Consolidate Mulitple Data Set Descriptions}
\usage{
consolidated_describe_data(...)
}
\arguments{
\item{...}{a set of formatted data.frames.}
}
\value{
A single data.frame composed of descriptions of all variables
for all data sets specified.
}
\description{
Consolidate Mulitple Data Set Descriptions
}
