% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe.r
\name{describe_data}
\alias{describe_data}
\title{Describe a data set}
\usage{
describe_data(x, data_name = NULL)
}
\arguments{
\item{x}{the data.frame with label attributes..}

\item{data_name}{the name of the data set to describe. If not null, then
a column, data_name is added and the value is repeated. Otherwise,
a new column is not added. Default NULL.}
}
\value{
A tibble with description information.
}
\description{
An ADaM-formatted .sas7bdat file includes a minimal amount
of information about each of the variables, which are stored as
attributes. This fuction "describes" these data sets and includes
the variable names, descriptions, types, and other summary information.
}
