% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dup_vars.r
\name{duplicated_vars}
\alias{duplicated_vars}
\title{Find Variables Appearing Multiple Data Sets}
\usage{
duplicated_vars(x, on, x_names = names(x))
}
\arguments{
\item{x}{the list of data.frames.}

\item{on}{the collapse variable.}

\item{x_names}{the names of the data sets. (Default names(x))}
}
\description{
Find Variables Appearing Multiple Data Sets
}
