% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove-equiv-columns.r
\name{remove_equiv_columns}
\alias{remove_equiv_columns}
\title{Remove redundant equivalent columns}
\usage{
remove_equiv_columns(x, keep_cols = character(), verbose = FALSE)
}
\arguments{
\item{x}{a data.frame that may have repeated, equivalent columns.}

\item{keep_cols}{a character vector of columsn that should be kept.}

\item{verbose}{should information about dropped columns be printed?
(default FALSE)}
}
\value{
a data frame where redundant columns have been dropeed.
}
\description{
Find the equivalant columns of a data.frame. Keep the first
remove the rest.
}
\examples{

iris$Sepal.Length2 <- 3 * iris$Sepal.Length + 3
remove_equiv_columns(iris)

}
